# generated-api-facade/openapi.yaml
openapi: 3.0.3
info:
  title: Generated Purchase Order API Facade
  version: 0.1.0
  description: >
    Demo API candidate for request_purchase_order_approval. UiPath should call
    this facade only after validation and trusted-tool registration approval.
servers:
  - url: http://localhost:8002
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service health
  /api/purchase-orders/{po_id}/approval-request:
    post:
      summary: Request purchase order approval
      parameters:
        - name: po_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequest"
      responses:
        "200":
          description: Approval request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
        "404":
          description: Purchase order not found
components:
  schemas:
    ApprovalRequest:
      type: object
      required:
        - approval_reason
        - manager_id
        - source_case_id
      properties:
        approval_reason:
          type: string
          example: Amount exceeds budget limit
        manager_id:
          type: string
          example: MGR-001
        source_case_id:
          type: string
          example: CASE-001
    ApprovalResponse:
      type: object
      required:
        - po_id
        - status
        - audit_log_created
        - execution_mode
        - source_case_id
      properties:
        po_id:
          type: string
          example: PO-1001
        status:
          type: string
          example: PENDING_MANAGER_APPROVAL
        audit_log_created:
          type: boolean
          example: true
        execution_mode:
          type: string
          example: API
        source_case_id:
          type: string
          example: CASE-001